requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.27
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
httpx==0.27.0
xgboost==2.0.3
shap==0.44.1

# Data Processing
pandas==2.2.0
numpy==1.26.4
scipy==1.12.0
scikit-learn==1.4.0
joblib==1.3.2

# Machine Learning
torch==2.2.0
torchvision==0.17.0

# Storage
pyarrow==15.0.0

# Security & Utils
slowapi==0.1.9
python-multipart==0.0.9

# LLM & Search
google-generativeai==0.8.3

# Fuzzy Name Matching
thefuzz==0.22.1
python-Levenshtein==0.25.1